requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.6
django-heroku==0.3.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyMySQL==1.0.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
urllib3==1.26.6
Werkzeug==2.0.1
whitenoise==5.3.0
